+++ crates/maple_cli/src/commands/grep.rs
@@ -18,7 +18,7 @@ use crate::commands::cache::{cache_exists, send_response_from_cache, SendRespons
 use crate::process::light::{set_current_dir, LightCommand};
 use crate::tools::ripgrep::Match;
 
-const RG_ARGS: [&str; 7] = [
+const RG_ARGS: &[&str] = &[
     "rg",
     "--column",
     "--line-number",
@@ -26,13 +26,12 @@ const RG_ARGS: [&str; 7] = [
     "--color=never",
     "--smart-case",
     "",
+    ".",
 ];
 
 // Ref https://github.com/liuchengxu/vim-clap/issues/533
-#[cfg(windows)]
+// Now `.` is pushed to the end for all platforms due to https://github.com/liuchengxu/vim-clap/issues/711.
 const RG_EXEC_CMD: &str = "rg --column --line-number --no-heading --color=never --smart-case '' .";
-#[cfg(not(windows))]
-const RG_EXEC_CMD: &str = "rg --column --line-number --no-heading --color=never --smart-case ''";
 
 #[derive(StructOpt, Debug, Clone)]
 pub struct Grep {